\name{hmmfb-package}
\alias{hmmfb-package}
\alias{hmmfb}
\docType{package}
\title{
\packageTitle{hmmfb}
}
\description{
\packageDescription{hmmfb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmmfb}
\packageIndices{hmmfb}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hmmfb}

Maintainer: \packageMaintainer{hmmfb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
